BAU-409 Move ErrorControlContext into common
