+++ src/explore-education-statistics-admin/src/hooks/useFormSubmit.ts
@@ -1,8 +1,8 @@
-import { useErrorControl } from '@admin/contexts/ErrorControlContext';
 import handleServerSideValidation, {
   ServerValidationErrors,
   ServerValidationMessageMapper,
 } from '@common/components/form/util/serverValidationHandler';
+import { useErrorControl } from '@common/contexts/ErrorControlContext';
 import { AxiosError } from 'axios';
 import { FormikActions } from 'formik';
 import { useMemo } from 'react';
