+++ src/explore-education-statistics-admin/src/pages/users/UserInvitePage.tsx
@@ -1,6 +1,5 @@
 import Link from '@admin/components/Link';
 import Page from '@admin/components/Page';
-import { ErrorControlState } from '@admin/contexts/ErrorControlContext';
 import useFormSubmit from '@admin/hooks/useFormSubmit';
 import userService from '@admin/services/users/service';
 import { UserInvite } from '@admin/services/users/types';
@@ -11,6 +10,7 @@ import Form from '@common/components/form/Form';
 import FormFieldTextInput from '@common/components/form/FormFieldTextInput';
 import { errorCodeToFieldError } from '@common/components/form/util/serverValidationHandler';
 import RelatedInformation from '@common/components/RelatedInformation';
+import { ErrorControlState } from '@common/contexts/ErrorControlContext';
 import Yup from '@common/lib/validation/yup';
 import React from 'react';
 import { RouteComponentProps } from 'react-router';
